Move code from .h to .cc in plugins.
